Two fragments connected via interface in activity
I created two fragments and placed them inside my main activity.  One
was a list fragment that contains two titles “thing 1”, and “thing 2”.
The second only contained a text view to show a corresponding
description.  Forgive me for my uncreative data.  I then connected both
fragments by passing the position of the chosen list view back to the
activity via an interface that I created.  Then from the main activity,
used the fragment manager to find the description fragment and set the
text of that fragment to what I wanted.  Fairly simple stuff and again
this is a proof of concept and me following along to Jim Wilson’s
Android for .NET Developers: 3 Adopting The Android Mindset that is
located on plural sight.